Changed adapter name to SearchAdapter Transferred read methods to adapter from fragment
